Rework graceful shutdown of service and include healthcheck endpoint
Gracful shutdown will first stop listening to any consumers groups
before coninuing to close connections to the server, other api's,
producing new messages to kafka before finally closing the consumer
group connection. This will aloow the service to commit an in flight
message on success in the given configurable time out, defaulted to 5
seconds.

Additionally a healthcheck endpoint has been added to the service and is
implemented with the graceful shutdown procedure.