Lower case dimension names for comparison
Each dimension name is matched to the corresponding dimension name from the V4 input file. The dimension name in the input file is lower cases before it is comparsed, but the dimension name from the instance is not. The dimension name from the instance originates from the recipe. If the recipe dimension names are updated to be capitalised, this process would break.