Separate file read from dimension post
Posting the dimensions to the dataset API as they are read from the file means the file reader is open for an unnecessary amount of time. For large datasets this increases the chance of getting connection reset errors from S3. Separating the file reading from the dimension posting allows the file to be read much quicker.